reduce file count, more progress
