--- Content.Server/Botany/Components/AgeGrowthComponent.cs
@@ -1,5 +1,0 @@
-namespace Content.Server.Botany.Components;
-
-[RegisterComponent]
-public sealed partial class AgeGrowthComponent : PlantGrowthComponent
-{}
