+++ Content.Server/Botany/Components/AtmosphericGrowthComponent.cs
@@ -1,8 +1,14 @@
 namespace Content.Server.Botany.Components;
 
 [RegisterComponent]
-public sealed partial class PressureGrowthComponent : PlantGrowthComponent
+public sealed partial class AtmosphericGrowthComponent : PlantGrowthComponent
 {
+    [DataField]
+    public float IdealHeat = 293f;
+
+    [DataField]
+    public float HeatTolerance = 10f;
+
     [DataField]
     public float LowPressureTolerance = 81f;
