--- Content.Server/Botany/Components/ConsumeGasGrowthComponent.cs
@@ -1,9 +1,0 @@
-using Content.Shared.Atmos;
-
-namespace Content.Server.Botany.Components;
-
-[RegisterComponent]
-public sealed partial class ConsumeGasGrowthComponent : PlantGrowthComponent
-{
-    [DataField] public Dictionary<Gas, float> ConsumeGasses = new();
-}
