--- Content.Server/Botany/Components/NutrientGrowthComponent.cs
@@ -1,8 +1,0 @@
-namespace Content.Server.Botany.Components;
-
-[RegisterComponent]
-public sealed partial class NutrientGrowthComponent : PlantGrowthComponent
-{
-    [DataField]
-    public float NutrientConsumption = 0.75f;
-}
