--- Content.Server/Botany/Components/TemperatureGrowthComponent.cs
@@ -1,11 +1,0 @@
-namespace Content.Server.Botany.Components;
-
-[RegisterComponent]
-public sealed partial class TemperatureGrowthComponent : PlantGrowthComponent
-{
-    [DataField]
-    public float IdealHeat = 293f;
-
-    [DataField]
-    public float HeatTolerance = 10f;
-}
