+++ Content.Server/Botany/Systems/AtmosphericGrowthSystem.cs
@@ -3,7 +3,7 @@ using Content.Server.Botany.Components;
 using Content.Shared.Atmos;
 
 namespace Content.Server.Botany.Systems;
-public sealed class PressureGrowthSystem : PlantGrowthSystem
+public sealed class AtmosphericGrowthSystem : PlantGrowthSystem
 {
     [Dependency] private readonly BotanySystem _botany = default!;
     [Dependency] private readonly PlantHolderSystem _plantHolderSystem = default!;
@@ -12,10 +12,10 @@ public sealed class PressureGrowthSystem : PlantGrowthSystem
     public override void Initialize()
     {
         base.Initialize();
-        SubscribeLocalEvent<PressureGrowthComponent, OnPlantGrowEvent>(OnPlantGrow);
+        SubscribeLocalEvent<AtmosphericGrowthComponent, OnPlantGrowEvent>(OnPlantGrow);
     }
 
-    private void OnPlantGrow(EntityUid uid, PressureGrowthComponent component, OnPlantGrowEvent args)
+    private void OnPlantGrow(EntityUid uid, AtmosphericGrowthComponent component, OnPlantGrowEvent args)
     {
         PlantHolderComponent? holder = null;
         Resolve<PlantHolderComponent>(uid, ref holder);
@@ -24,6 +24,18 @@ public sealed class PressureGrowthSystem : PlantGrowthSystem
             return;
 
         var environment = _atmosphere.GetContainingMixture(uid, true, true) ?? GasMixture.SpaceGas;
+        if (MathF.Abs(environment.Temperature - component.IdealHeat) > component.HeatTolerance)
+        {
+            holder.Health -= _random.Next(1, 3);
+            holder.ImproperHeat = true;
+            if (holder.DrawWarnings)
+                holder.UpdateSpriteAfterUpdate = true;
+        }
+        else
+        {
+            holder.ImproperHeat = false;
+        }
+
         var pressure = environment.Pressure;
         if (pressure < component.LowPressureTolerance || pressure > component.HighPressureTolerance)
         {
