--- Content.Server/Botany/Systems/NutrientGrowthSystem.cs
@@ -1,44 +1,0 @@
-using Content.Server.Botany.Components;
-using Robust.Shared.Random;
-
-namespace Content.Server.Botany.Systems;
-public sealed class NutrientGrowthSystem : PlantGrowthSystem
-{
-    public override void Initialize()
-    {
-        base.Initialize();
-        SubscribeLocalEvent<NutrientGrowthComponent, OnPlantGrowEvent>(OnPlantGrow);
-    }
-
-    private void OnPlantGrow(EntityUid uid, NutrientGrowthComponent component, OnPlantGrowEvent args)
-    {
-        PlantHolderComponent? holder = null;
-        Resolve<PlantHolderComponent>(uid, ref holder);
-
-        if (holder == null || holder.Seed == null || holder.Dead)
-            return;
-
-        if (component.NutrientConsumption > 0 && holder.NutritionLevel > 0 && _random.Prob(0.75f))
-        {
-            holder.NutritionLevel -= MathF.Max(0f,
-                component.NutrientConsumption * HydroponicsConsumptionMultiplier * HydroponicsSpeedMultiplier);
-            if (holder.DrawWarnings)
-                holder.UpdateSpriteAfterUpdate = true;
-        }
-
-        var healthMod = _random.Next(1, 3) * HydroponicsSpeedMultiplier;
-        if (holder.SkipAging < 10)
-        {
-            // Make sure the plant is not hungry
-            if (holder.NutritionLevel > 5)
-            {
-                holder.Health += Convert.ToInt32(_random.Prob(0.35f)) * healthMod;
-            }
-            else
-            {
-                AffectGrowth(-1, holder);
-                holder.Health -= healthMod;
-            }
-        }
-    }
-}
