--- Content.Server/Botany/Systems/WaterGrowthSystem.cs
@@ -1,44 +1,0 @@
-using Content.Server.Botany.Components;
-using Robust.Shared.Random;
-
-namespace Content.Server.Botany.Systems;
-public sealed class WaterGrowthSystem : PlantGrowthSystem
-{
-    public override void Initialize()
-    {
-        base.Initialize();
-        SubscribeLocalEvent<WaterGrowthComponent, OnPlantGrowEvent>(OnPlantGrow);
-    }
-
-    private void OnPlantGrow(EntityUid uid, WaterGrowthComponent component, OnPlantGrowEvent args)
-    {
-        PlantHolderComponent? holder = null;
-        Resolve<PlantHolderComponent>(uid, ref holder);
-
-        if (holder == null || holder.Seed == null || holder.Dead)
-            return;
-
-        if (component.WaterConsumption > 0 && holder.WaterLevel > 0 && _random.Prob(0.75f))
-        {
-            holder.WaterLevel -= MathF.Max(0f,
-                component.WaterConsumption * HydroponicsConsumptionMultiplier * HydroponicsSpeedMultiplier);
-            if (holder.DrawWarnings)
-                holder.UpdateSpriteAfterUpdate = true;
-        }
-
-        var healthMod = _random.Next(1, 3) * HydroponicsSpeedMultiplier;
-        if (holder.SkipAging < 10)
-        {
-            // Make sure the plant is not thirsty.
-            if (holder.WaterLevel > 10)
-            {
-                holder.Health += Convert.ToInt32(_random.Prob(0.35f)) * healthMod;
-            }
-            else
-            {
-                AffectGrowth(-1, holder);
-                holder.Health -= healthMod;
-            }
-        }
-    }
-}
